requirements.txt
python-dotenv==0.21.0  
pandas    
requests==2.28.1       
psycopg2>=2.9.5 
sqlalchemy==2.0.0
numpy>=1.20      